feat: reorganize entry points
