+++ package.json
@@ -45,11 +45,13 @@
     "build": "concurrently -m 8 'yarn:build:*'",
     "build-watch": "concurrently -n watch: 'yarn:build:* --watch'",
     "build:base": "rollup -c",
-    "build:vanilla": "rollup -c --config-vanilla",
     "build:utils": "rollup -c --config-utils",
-    "build:react-utils": "rollup -c --config-react-utils",
+    "build:vanilla": "rollup -c --config-vanilla",
+    "build:vanilla_utils": "rollup -c --config-vanilla_utils",
+    "build:react": "rollup -c --config-react",
+    "build:react_utils": "rollup -c --config-react_utils",
     "build:macro": "rollup -c --config-macro",
-    "build:macro-vite": "rollup -c --config-macro_vite",
+    "build:macro_vite": "rollup -c --config-macro_vite",
     "postbuild": "yarn copy && yarn patch-macro-vite && yarn patch-ts3.4 && yarn patch-esm-ts",
     "prettier": "prettier '*.{js,json,md}' '{src,tests,docs}/**/*.{ts,tsx,md,mdx}' --write",
     "prettier:ci": "prettier '*.{js,json,md}' '{src,tests,docs}/**/*.{ts,tsx,md,mdx}' --list-different",
