+++ src/index.ts
@@ -1,9 +1,2 @@
-export { useSnapshot } from './react'
-export {
-  ref,
-  proxy,
-  getVersion,
-  subscribe,
-  snapshot,
-  unstable_buildProxyFunction,
-} from './vanilla'
+export * from './vanilla'
+export * from './react'
