--- src/react-utils.ts
@@ -1,1 +1,0 @@
-export { useProxy } from './react-utils/useProxy'
