+++ src/react/utils.ts
@@ -1,0 +1,1 @@
+export { useProxy } from './utils/useProxy'
