+++ src/utils.ts
@@ -1,9 +1,2 @@
-export { subscribeKey } from './utils/subscribeKey'
-export { watch } from './utils/watch'
-export { devtools } from './utils/devtools'
-export { derive, underive, unstable_deriveSubscriptions } from './utils/derive'
-export { addComputed_DEPRECATED as addComputed } from './utils/addComputed'
-export { proxyWithComputed } from './utils/proxyWithComputed'
-export { proxyWithHistory } from './utils/proxyWithHistory'
-export { proxySet } from './utils/proxySet'
-export { proxyMap } from './utils/proxyMap'
+export * from './vanilla/utils'
+export * from './react/utils'
